Rename as macro-catch-all middleware
